scss is hard :(
spent most of the day learning the scss we have styling the site

once I understood how it worked I implement styling for small phones as well as regular sized phone. The only catch was the images in the project cards. I couldnt find a way to style it nicely with how our markup is currenty set up. I will need some pr css god to help me with that one for sure.